♻️ Refactor: move different trick implementation to different files
