+++ pkg/threepilestrick/trick.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	"github.com/carloscasalar/card-guess/internal/deck"
-	"github.com/carloscasalar/card-guess/internal/mat"
 )
 
 type Card string
@@ -39,95 +38,11 @@ func New(shuffleBeforeInitialDraw bool) (Trick, error) {
 	if err != nil {
 		return nil, err
 	}
-	return trickState{
+	return intermediateTrickState{
 		sample:          sample,
 		mat:             *theMat,
 		pileChooseCount: 0,
 	}, nil
-}
-
-type trickState struct {
-	sample          deck.Pile
-	mat             mat.Mat
-	pileChooseCount int
-}
-
-func (i trickState) GuessMyCard() (*Card, error) {
-	return nil, ErrStillCannotGuessYourCard
-}
-
-func (i trickState) Sample() Pile {
-	return pileToSerializablePile(i.sample)
-}
-
-func (i trickState) FirstPile() Pile {
-	return pileToSerializablePile(i.mat.FirstPile())
-}
-
-func (i trickState) SecondPile() Pile {
-	return pileToSerializablePile(i.mat.SecondPile())
-}
-
-func (i trickState) ThirdPile() Pile {
-	return pileToSerializablePile(i.mat.ThirdPile())
-}
-
-func (i trickState) MyCardIsInPile(holder PileHolder) (Trick, error) {
-	const requiredChooseRounds = 3
-	pileChooseCount := i.pileChooseCount + 1
-	chosenPileHolder := mat.PileHolder(holder)
-	if pileChooseCount == requiredChooseRounds {
-		return &finalTrickState{sample: i.sample, mat: i.mat, pileWhereChosenCardIs: chosenPileHolder}, nil
-	}
-	allCardsWithChosenPileInTheMiddle := i.mat.JoinWithPileInTheMiddle(chosenPileHolder)
-	newMat, err := splitIntoThreePiles(allCardsWithChosenPileInTheMiddle)
-	if err != nil {
-		return nil, err
-	}
-	return trickState{
-		sample:          i.sample,
-		mat:             *newMat,
-		pileChooseCount: pileChooseCount,
-	}, nil
-}
-
-type finalTrickState struct {
-	sample                deck.Pile
-	mat                   mat.Mat
-	pileWhereChosenCardIs mat.PileHolder
-}
-
-func (f finalTrickState) Sample() Pile {
-	return pileToSerializablePile(f.sample)
-}
-
-func (f finalTrickState) FirstPile() Pile {
-	return pileToSerializablePile(f.mat.FirstPile())
-}
-
-func (f finalTrickState) SecondPile() Pile {
-	return pileToSerializablePile(f.mat.SecondPile())
-}
-
-func (f finalTrickState) ThirdPile() Pile {
-	return pileToSerializablePile(f.mat.ThirdPile())
-}
-
-func (f finalTrickState) GuessMyCard() (*Card, error) {
-	const positionWhereCardIsAfterThreeSplits = 4
-	card, err := f.mat.CardFromPile(f.pileWhereChosenCardIs, positionWhereCardIsAfterThreeSplits)
-	if err != nil {
-		return nil, err
-	}
-	theCard := Card(card.String())
-	return &theCard, nil
-}
-
-func (f finalTrickState) MyCardIsInPile(holder PileHolder) (Trick, error) {
-	if mat.PileHolder(holder) != f.pileWhereChosenCardIs {
-		return nil, ErrAskMeToGuessTheCardInstead
-	}
-	return f, nil
 }
 
 func pileToSerializablePile(pile deck.Pile) Pile {
